added libra_partition.cc and moved libra code from kernel.cc to libra_partition.cc
